fps 토글기능 추가, 레이더 오류 수정, 해상도/창모드 옵션 추가, 발소리 오류 수정
전체화면 했을 때의 fps를 확인하기 위한 ui와, 레이더에 더미 오브젝트가 보이던 오류, 상대 플레이어의 발소리가 제대로 나지 않던 오류를 수정하고 발소리의 거리 배율을 늘렸으며, 처음 실행시 다이얼로그에서 해상도 및 창모드 여부를 설정할 수 있도록 수정하였다.

+++ AvoidHimClient/Camera.cpp
@@ -5,7 +5,7 @@
 Camera::Camera() {
 	viewTransform = Matrix4x4::Identity();
 	projectionTransform = Matrix4x4::Identity();
-	viewPort = { 0,0, C_WIDTH, C_HEIGHT, 0, 1 };
+	viewPort = { 0,0, (float)C_WIDTH, (float)C_HEIGHT, 0, 1 };
 	scissorRect = { 0,0, C_WIDTH, C_HEIGHT };
 
 	localOffset = XMFLOAT3(0, 0, 0);
